+++ src/main/java/com/athenhub/stockservice/stock/infrastructure/client/ProductClient.java
@@ -1,6 +1,6 @@
 package com.athenhub.stockservice.stock.infrastructure.client;
 
-import com.athenhub.stockservice.stock.infrastructure.config.FeignClientConfig;
+import com.athenhub.stockservice.global.infrastructure.feignclient.FeignClientConfig;
 import java.util.UUID;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
